Line-up interfaces to use size_t for GetCount()s.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@38019 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/richtext/richtextbuffer.h
@@ -1492,7 +1492,7 @@ public:
     virtual void ClearStyleStack();
 
     /// Get the size of the style stack, for example to check correct nesting
-    virtual int GetStyleStackSize() const { return m_attributeStack.GetCount(); }
+    virtual size_t GetStyleStackSize() const { return m_attributeStack.GetCount(); }
 
     /// Begin using bold
     bool BeginBold();
